# docs/swagger.yaml
basePath: /
definitions:
...
  internal_interfaces_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_interfaces_http.GenericResponse:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
...
info:
  contact: {}
  description: Internal deployment platform API.
  title: GOS Release API
  version: "1.0"
paths:
...
  /permissions:
    get:
      description: 查询Permissions列表，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询Permissions列表
      tags:
      - users
...
  /users:
    get:
      description: 查询Users列表，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询Users列表
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 创建User，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 创建User
      tags:
      - users
  /users/{id}:
    delete:
      description: 删除User，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 删除User
      tags:
      - users
    get:
      description: 获取User By ID详情，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 获取User By ID详情
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 更新User，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 更新User
      tags:
      - users
  /users/{id}/param-permissions:
    get:
      description: 查询User Param Permissions列表，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询User Param Permissions列表
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 处理Upsert User Param Permission接口，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 处理Upsert User Param Permission接口
      tags:
      - users
...
  /users/options:
    get:
      description: 查询User Options列表，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询User Options列表
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"